requirements.txt
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3